Concurrency fixes (OTH-*) + stopped-state docs/decisions (CER-1, SLB-3)
Remaining concurrency/correctness items from the stoppable-pattern
review, plus the two decide-and-document items:

- OTH-1: SlabAllocator uses one condition variable per wait-predicate
  (free_cv / init_cv / size_cv / empty_cv) instead of one shared cv.
  This fixes the lost-wakeup hazard where return_slab()'s notify_one
  could wake a waiter with a different predicate (e.g. a
  block_until_empty() caller) while a get_slab(blocking=true) waiter
  slept forever. notify_one on free_cv is now provably sound: every
  waiter shares the same predicate and one returned slab satisfies
  exactly one of them. stop() notify_all's all four.

- OTH-2: FakeXEngine::_read_acks notifies w.cv when it drains
  ack_queue. Fixes a debug-mode hang: the worker's idle-loop ack drain
  popped the last ack without any notify, leaving a synchronize()
  caller (predicate: command_queue and ack_queue both empty) asleep
  forever.

- OTH-3: AssembledFrameAllocator nfreq / beam_ids are read through new
  lock-synchronized getters (get_nfreq / get_beam_ids) in the pybind11
  bindings, instead of raw def_readonly reads of lock-protected members
  (torn vector copy in the worst case). time_samples_per_chunk stays
  def_readonly (ctor-constant).

- OTH-5: thread-handle publication races fixed in four classes, all by
  the same recipe -- publish under the mutex, synchronize the join path:
  (a) Hwtest 'threads' published under the mutex in start(); new
  _join_threads() helper (used by join() and the dtor) takes the mutex
  briefly before joining. (b) HwtestSender::start() holds the lock
  across the worker spawns; join() synchronizes likewise. (c)
  FrbGrouper::open() publishes grpc_state->server and send_thread under
  the mutex (which also lets the send_io_done failure-path catch drop
  its inner lock); close() is serialized by a new leaf-level
  close_mutex, so a second closer blocks until teardown completes
  instead of returning early into a potential use-after-free (e.g.
  destructor racing an in-flight close()). (d) FrbServer::start() holds
  the mutex through rpc_server and all thread spawns; stop()/dtor
  already acquire it first. In each class, holding the lock across the
  spawns is deadlock-free: freshly spawned threads block briefly on the
  mutex, and the spawner never waits on them.

- OTH-6: AssembledFrameAllocator::_create_frame_set resets
  frame_creation_underway via a dismissable scope guard if it unwinds
  on a throw (re-acquiring the lock if the throw came from the unlocked
  middle section). Previously the flag stayed set forever, parking all
  later get_frame_set() callers -- masked only by the entry-point
  wrapper's stop() coupling.

- OTH-7: Barrier::wait_count (generation counter) int -> long, per
  cpp.md; signed overflow was UB and a potential predicate hang in a
  long-running server.

- OTH-8: SlabAllocator dummy path passes the local aligned_nbytes to
  _af_alloc instead of re-reading slab_size after unlock (formal data
  race).

- OTH-10: HwtestSender::worker_main moves endpoints.at() and the
  exit-summary print inside a catch-all so a pathological throw cannot
  escape the thread (the exit counter stays outside -- wait() needs it
  on every path); BumpAllocator::_zero_worker's last-out-finalizes
  block gets its own try/catch, since it must still run after a caught
  zeroing error.

- CER-1 (doc-only): the blocking_sync note in CudaEventRingbuf.hpp now
  documents the uninterruptible cudaEventSynchronize trade-off: the
  wait is bounded by the recorded-events-only invariant unless the GPU
  itself hangs, in which case stop()/destructors hang; a
  cudaEventQuery-polling mode is noted as the escape hatch. Matching
  accepted-risk comment at ~GpuDedisperser's cudaStreamSynchronize loop.

- SLB-3: SlabAllocator stopped-state semantics decided and documented.
  wait_until_initialized() now checks the allocator's own stopped state
  first, making the behavior uniform across modes (previously an
  error-stopped dummy/aflags-mode allocator silently succeeded). The
  header classifies entry points (get_slab, block_until_empty,
  num_total_slabs, wait_until_initialized -- methods that can block)
  vs stopped-tolerant informational accessors (num_free_slabs,
  get_slab_size, is_initialized, is_dummy), with per-method notes and
  deliberate no-stopped-check comments in the accessor bodies.

Verified: make -j 32 clean; full suite 'pirate_frb test -n 2' passed on
the OTH batch, and 'test --net --serv' passed on top of CER-1 + SLB-3.

+++ include/pirate/Barrier.hpp
@@ -24,10 +24,12 @@ struct Barrier
     mutable bool is_stopped = false;
     mutable std::exception_ptr error;
 
-    // Protected by lock.
+    // Protected by lock. wait_count is a generation counter (one increment
+    // per barrier release); 'long' so it cannot overflow (UB, and a hung
+    // 'wait_count > wc' predicate) in a long-running server.
     int nthreads = 0;
     int nthreads_waiting = 0;
-    int wait_count = 0;
+    long wait_count = 0;
 
     // If constructor is called with nthreads=0, then 'nthreads' must be
     // set later, with a call to initialize().
